fixing a crash in 1.11.2

+++ src/main/java/omtteam/omlib/util/compat/MathTools.java
@@ -37,7 +37,7 @@ public class MathTools {
         return MathHelper.absMax(max1, max2);
     }
 
-    public static double sqrt_double(double sqrt) {
+    public static float sqrt_double(double sqrt) {
         return MathHelper.sqrt(sqrt);
     }
